+++ core/datastore/src/main/java/com/jparkbro/datastore/UserDataStore.kt
@@ -1,8 +1,8 @@
 package com.jparkbro.datastore
 
 interface UserDataStore {
-    suspend fun saveUserInfo(userId: Int, nickname: String): Result<Unit>
-    suspend fun getUserId(): Result<Int>
+    suspend fun saveUserInfo(userId: Long, nickname: String): Result<Unit>
+    suspend fun getUserId(): Result<Long>
     suspend fun getUserNickName(): Result<String>
 
     suspend fun clearUserInfo(): Result<Unit>
